backend/requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
ecdsa==0.17.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
iso8601==1.0.2
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.9.1
pypika-tortoise==0.1.5
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
requests==2.28.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tortoise-orm==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
